Changed VERSION to be a string using major.minor To be more compliant with typical semantic versioning the version in the VERSION file is changed to major.minor.
The major version number indicates changes to VSS that are not
backwards compatible such as deleted signals, renamed signals, etc.
In this case a new ID file is also created.

The minor version number indicates changes to VSS that are backwards
compatible such as new signals being added (the ID file is amended),
changed descriptions.

We are not using a patchlevel as it does not apply.

Signed-off-by: Rudolf J Streif <rstreif@jaguarlandrover.com>