Start git subtree separation from main VSS tree
This is an empty commit containing only this information commit message.
The parent of this commit is the start of a git subtree split,
containing the content of the tools/ directory (and its related history)
split off from the vehicle_signal_specification (VSS) repository into
this independent repository.

The intention is to allow the tools development to work independently,
and keep separate git histories for the two aspects (signal database
content, vs. software related to it).  This repository may however be
included for convenience into VSS as a git submodule.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>