Replace vspec2franca with new version in contrib
The previous version created a single type named SignalSpec which was
a single array of signal definitions.   This has little meaning other
than a similar "list" in CSV or JSON - in other words could mostly just
be read and processed further as input.

This version more correctly translates VSS into one or several Franca
Interfaces, containing one Attribute per VSS signal.  Franca Attributes
(a.k.a. properties in other languages), are data items that can be
observed / subscribed-to for changes.  This semantically matches a VSS
signal, and through existing tools it is possible to translate this
directly to a protocol binding.  For example in SOME/IP it would expose
the named signal as a "field" in SOME/IP that it can be observed
(subscribe for change notifications using SOME/IPs built in mechanism
for this).  It is similar in other protocols.

This version uses the .csv file as input in order to process already
pre-processed tree names, including expansion of instances.
A future version might rather be based directly on vspec.py, and could
possibly make use of the instance concept to define a base
class/interface, and inherited class/interfaces.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>