Generate a deployment file (.fdepl) for CommonAPI C++ SOME/IP
Deployment files are optionally used in some cases to provide more
information to the code-generation binding.  For SOME/IP, it is
mandatory to provide one, which includes defining exact numerical IDs
for each service and each callable methods/attribute.  We use a
random-number (within range) strategy to create these IDs.

BUG! Note, this strategy is naive and not performant for a large amount
of signals/attributes in one go.  The required time to find a unique
random number will go up exponentially and may ultimately fail. An
improvement to this should be made for production level use.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>