Use /usr/bin/env python3 everywhere
According to what I have learned, Linux distros are recommended to ship
the 'python3' alias even if /usr/bin/python is already version 3.  It
should be safe for quite some time to force the use of python3 this way.

In addition, by using the env wrapper, if a virtual environment has been
set up then that version will be used instead of forcing the system
installed version /usr/bin/python[3].  This can be useful when testing
in CI.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>