setup.py: Tools version & name independent of VSS
Previous script used the VERSION file in the VSS main repository
to set the tools version but since vss-tools is now a separate
repository (submodule), it should be versioned separately and should
also not be strictly depending on its parent.

The strategy is to use the tag name if it is a tagged version, and
otherwise use a 8-character abbreviated commit hash.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>