Move ocf related tools and create contrib/ dir
Introducing contrib/ dir as was proposed on a recent VSS meeting.

The idea is to here keep tools that are still being evaluated,
work-in-progress, or otherwise not guaranteed to be maintained,
but still interesting enough to keep or introduce already now.

The OCF tooling has not been used or updated for a long time so it
should not be in the primary directory at the moment.