fix python packaging and file structure
move `vspec.py` into `vspec/__init__.py` and `model/` inside it, also
with an `__init__.py` to be handled as a Python package.

This allows proper installation and avoid conflicts with `model` being
a toplevel module, which could conflict with other projects.

The existing files were changed to use the new `model` location,
inside the `vspec` package.

Other changes:
 - Python >= 3.7, matching the README.md
 - added pyproject.toml as per https://setuptools.readthedocs.io/
 - declared install_requires and tests_require dependencies