Renaming exclude_private to merge_private
The parameter exclude_private in load_tree() is being passed to the
render_tree() function as merge_private. This behavior can confuse
the programmer since both names mean opposite ideas. The real behavior
is that the variable merges Private if it's true, so this commit
corrects this bug just by changing the variable name.

Resolves #85